move `categorize()` to `utils.py` and make `set_parser_prog` a private method
